% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday.R
\name{holiday_countries}
\alias{holiday_countries}
\title{Fetch list of supported countries}
\usage{
holiday_countries(language = "EN")
}
\arguments{
\item{language}{Language for country names (default "EN")}
}
\value{
tibble: iso_code, name, official_languages
}
\description{
Fetch list of supported countries
}
