% Generated by build script
\name{holiday_countries}
\alias{holiday_countries}
\title{Fetch list of supported countries}
\usage{
holiday_countries()
}
\arguments{
\item{language}{Language for country names (default "EN")}
}
\value{tibble: iso_code, name, official_languages}
\description{Fetch list of supported countries}
