% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday.R
\name{holiday_public}
\alias{holiday_public}
\title{Fetch public holidays for a country and year}
\usage{
holiday_public(country, year = format(Sys.Date(), "\%Y"), language = "EN")
}
\arguments{
\item{country}{ISO 3166-1 alpha-2 country code (e.g. "US", "DE", "FR")}

\item{year}{Year (default current year)}

\item{language}{Language for holiday names (default "EN")}
}
\value{
tibble: id, start_date, end_date, type, name, nationwide
}
\description{
Fetch public holidays for a country and year
}
